refactor : chat 관련 controller 폴더 위치 변경
